+++ composeApp/src/commonTest/kotlin/dev/brewkits/kmpworkmanager/sample/ComposeAppCommonTest.kt
@@ -1,0 +1,12 @@
+package dev.brewkits.kmpworkmanager.sample
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ComposeAppCommonTest {
+
+    @Test
+    fun example() {
+        assertEquals(3, 1 + 2)
+    }
+}